+++ src/be_atom_memo_tuple5_/index.d.ts
@@ -2,32 +2,36 @@ import type { WritableAtom_ } from '@ctx-core/nanostores'
 import type { Be, be__params_T, Ctx } from '@ctx-core/object'
 import type { WritableAtom } from 'nanostores'
 export declare function be_atom_memo_tuple5_<
-	T = any
->():be_atom_memo_tuple5_T<T, WritableAtom_<T>>
+	T = any,
+	A extends WritableAtom<T> = WritableAtom_<T>
+>():be_atom_memo_tuple5_T<T, A>
 export declare function be_atom_memo_tuple5_<
-	T = any
+	T = any,
+	A extends WritableAtom<T> = WritableAtom_<T>
 >(
 	be__params?:be__params_T
-):be_atom_memo_tuple5_T<T, WritableAtom_<T>>
+):be_atom_memo_tuple5_T<T, A>
 export declare function be_atom_memo_tuple5_<
-	T = any
->():be_atom_memo_tuple5_T<T, WritableAtom_<T>>
+	T = any,
+	A extends WritableAtom<T> = WritableAtom_<T>
+>():be_atom_memo_tuple5_T<T, A>
 export declare function be_atom_memo_tuple5_<
-	T = any
+	T = any,
+	A extends WritableAtom<T> = WritableAtom_<T>
 >(
 	name:string|null|undefined,
 	be__params?:be__params_T
-):be_atom_memo_tuple5_T<T, WritableAtom_<T>>
+):be_atom_memo_tuple5_T<T, A>
 export declare function be_atom_memo_tuple5_<
 	T = any,
-	A extends WritableAtom<T> = WritableAtom<T>
+	A extends WritableAtom<T> = WritableAtom_<T>
 >(
 	atom__new:(ctx:Ctx)=>A,
 	be__params?:be__params_T
 ):be_atom_memo_tuple5_T<T, A>
 export declare function be_atom_memo_tuple5_<
 	T = any,
-	A extends WritableAtom<T> = WritableAtom<T>
+	A extends WritableAtom<T> = WritableAtom_<T>
 >(
 	name?:string|null|undefined,
 	atom__new?:((ctx:Ctx)=>A),
@@ -35,7 +39,7 @@ export declare function be_atom_memo_tuple5_<
 ):be_atom_memo_tuple5_T<T, A>
 export type be_atom_memo_tuple5_T<
 	T = any,
-	A extends WritableAtom<T> = WritableAtom<T>
+	A extends WritableAtom<T> = WritableAtom_<T>
 > = [
 	Be<A>,
 	(ctx:Ctx)=>T,